+++ src/renderer/src/components/message/MessageBlock.tsx
@@ -3,6 +3,7 @@ import {
   HELPERS_TOOL_GROUP_NAME,
   HELPERS_TOOL_INVALID_TOOL_ARGUMENTS,
   HELPERS_TOOL_NO_SUCH_TOOL,
+  POWER_TOOL_AGENT,
   POWER_TOOL_FILE_EDIT,
   POWER_TOOL_FILE_WRITE,
   POWER_TOOL_GROUP_NAME,
@@ -30,6 +31,7 @@ import {
 import { ToolMessageBlock } from '@/components/message/ToolMessageBlock';
 import { FileWriteToolMessage } from '@/components/message/FileWriteToolMessage';
 import { EditFileToolMessage } from '@/components/message/EditFileToolMessage';
+import { AgentToolMessage } from '@/components/message/AgentToolMessage';
 
 type Props = {
   baseDir: string;
@@ -79,6 +81,9 @@ export const MessageBlock = ({ baseDir, message, allFiles, renderMarkdown, remov
       }
       if (toolMessage.toolName === POWER_TOOL_FILE_EDIT) {
         return <EditFileToolMessage message={toolMessage} onRemove={remove} />;
+      }
+      if (toolMessage.toolName === POWER_TOOL_AGENT) {
+        return <AgentToolMessage message={toolMessage} onRemove={remove} />;
       }
     }